# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: registrations
    description: Course registration operations
  - name: auth
    description: Student authentication operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /registrations:
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Create a course registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Invalid registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginStudent
      tags: [auth]
      summary: Log in a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request a password reset link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List course registration orders for review
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, pending, paid, rejected]
            default: all
      responses:
        "200":
          description: Registration orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminOrder"
  /admin/orders/{id}/review:
    patch:
      operationId: reviewAdminOrder
      tags: [admin]
      summary: Approve or reject a registration order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderReviewInput"
      responses:
        "200":
          description: Updated registration order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrder"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/summary:
    get:
      operationId: getAdminSummary
      tags: [admin]
      summary: Get registration summary
      responses:
        "200":
          description: Registration summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSummary"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegistrationInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        receiptUrl:
          type: string
          minLength: 1
        consent:
          type: boolean
      required: [fullName, phone, email, password, receiptUrl, consent]
    Registration:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [pending, paid, rejected]
        accountStatus:
          type: string
          enum: [pending, active, inactive]
        createdAt:
          type: string
          format: date-time
      required: [id, fullName, phone, email, status, accountStatus, createdAt]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [email, password]
    LoginResult:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, inactive]
        message:
          type: string
      required: [status, message]
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    AdminOrder:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        receiptUrl:
          type: string
        status:
          type: string
          enum: [pending, paid, rejected]
        accountStatus:
          type: string
          enum: [pending, active, inactive]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, fullName, phone, email, receiptUrl, status, accountStatus, rejectionReason, createdAt]
    OrderReviewInput:
      type: object
      properties:
        decision:
          type: string
          enum: [approve, reject]
        reason:
          type: string
      required: [decision]
    AdminSummary:
      type: object
      properties:
        totalRegistrations:
          type: integer
        pendingOrders:
          type: integer
        approvedOrders:
          type: integer
        totalRevenue:
          type: integer
      required: [totalRegistrations, pendingOrders, approvedOrders, totalRevenue]