# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 8 */
  phone: string;
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  receiptUrl: string;
  consent: boolean;
}